backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.7.4
numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==10.4.0
onnxruntime==1.17.3
scikit-image==0.22.0
python-dotenv==1.0.1
